refactor: fix component name
