+++ src/constants/keys.js
@@ -32,7 +32,7 @@ export default {
     HEAT_MAP: 'heat-map',
     HEAT_MAP_TOOLTIP: 'heat-map-tooltip',
     HEAT_MAP_LABELS: 'heat-map-labels',
-    HEAT_MAP_HIGH_LIGHT: 'heat-map-high-light',
+    HEAT_MAP_HIGHLIGHT: 'heat-map-highlight',
     MINI_CHART_BACKGROUND: 'mini-chart-background-rect',
     MINI_CHART_NAVIGATION: 'mini-chart-navigation-rect',
     MINI_CHART_POINT: 'mini-chart-points',
